Consolidate optimizer enumeration and split bounds module
Extract shared pool prep and five-slot DFS into src/lib/optimizer/enumeration/
so search and combo-count share one code path. Move estimateOptimizerComboCount
to combo-count.ts to break the bounds/combo-count circular import, and split
bounds.ts into focused modules (independent, heuristic, joint).

Also debounce stat slider bounds separately from search (150ms), guard
searchLoadouts when no stat targets or set bonuses, remove dead worker cancel
message handling, and add npm run test:unit.
