+++ src/lib/optimizer/combo-count.ts
@@ -1,19 +1,22 @@
 import { SLOT_ORDER } from "@/lib/bungie/constants";
-import { ARMOR_STAT_NAMES, type ArmorStatName, type DerivedArmorPieceJson } from "@/lib/db/types";
-import { getPieceStatCeiling } from "@/lib/inventory/compute-stat-totals";
-import { estimateOptimizerComboCount } from "@/lib/optimizer/bounds";
+import {
+  ARMOR_STAT_NAMES,
+  type ArmorStatName,
+  type DerivedArmorPieceJson,
+} from "@/lib/db/types";
+import { dedupeSlotPieces } from "@/lib/optimizer/dedupe";
+import { enumerateLoadouts } from "@/lib/optimizer/enumeration/enumerate-loadouts";
+import {
+  groupPoolBySlot,
+} from "@/lib/optimizer/enumeration/pool-by-slot";
+import { prepareDedupedSlotPool } from "@/lib/optimizer/enumeration/prepare-slot-pool";
 import {
   hasStatTargets,
-  partialCanReachMins,
   totalsFromPieces,
 } from "@/lib/optimizer/constraints";
-import { dedupeSlotPieces } from "@/lib/optimizer/dedupe";
 import {
   applyExoticLockToSlotGroups,
-  countExoticsInPieces,
   DEFAULT_EXOTIC_LOCK,
-  exoticAllowedInPartialCombo,
-  resolveLockedExoticIdentityKey,
   type ExoticLock,
 } from "@/lib/optimizer/exotic-lock";
 import { addStatOffsets } from "@/lib/optimizer/fragment-offset";
@@ -23,7 +26,6 @@ import {
 } from "@/lib/optimizer/mod-offset";
 import { resolveLoadoutTotals } from "@/lib/optimizer/resolve-loadout-totals";
 import {
-  partialCanSatisfySetBonuses,
   satisfiesSetBonuses,
   type SetBonusSelection,
 } from "@/lib/optimizer/set-bonus";
@@ -34,34 +36,73 @@ import {
   clampOptimizerStat,
 } from "@/lib/optimizer/stat-range";
 
-function groupPoolBySlot(pool: DerivedArmorPieceJson[]) {
-  const bySlot = new Map<
+function dedupedRepresentativeCount(pieces: DerivedArmorPieceJson[]): number {
+  if (pieces.length === 0) return 0;
+  return dedupeSlotPieces(pieces).representatives.length;
+}
+
+function productSlotCounts(
+  slotPieces: Map<
+    DerivedArmorPieceJson["slot"],
+    DerivedArmorPieceJson[]
+  >,
+): number {
+  let product = 1;
+  for (const slot of SLOT_ORDER) {
+    const count = dedupedRepresentativeCount(slotPieces.get(slot) ?? []);
+    if (count === 0) return 0;
+    product *= count;
+  }
+  return product;
+}
+
+/**
+ * Deduped five-slot combination count (same shaping as search / joint bounds).
+ * Respects exotic lock rules: none = no exotics, locked = one pinned exotic,
+ * any = at most one exotic anywhere in the loadout (not one per slot).
+ */
+export function estimateOptimizerComboCount(
+  pool: DerivedArmorPieceJson[],
+  exoticLock: ExoticLock = DEFAULT_EXOTIC_LOCK,
+): number {
+  const bySlot = groupPoolBySlot(pool);
+
+  if (exoticLock.mode === "none" || exoticLock.mode === "locked") {
+    applyExoticLockToSlotGroups(bySlot, exoticLock, pool);
+    return productSlotCounts(bySlot);
+  }
+
+  const legendariesBySlot = new Map<
     DerivedArmorPieceJson["slot"],
     DerivedArmorPieceJson[]
   >();
   for (const slot of SLOT_ORDER) {
-    bySlot.set(slot, []);
+    legendariesBySlot.set(
+      slot,
+      (bySlot.get(slot) ?? []).filter((piece) => !piece.isExotic),
+    );
   }
-  for (const piece of pool) {
-    bySlot.get(piece.slot)?.push(piece);
-  }
-  return bySlot;
-}
 
-function perSlotMaxima(
-  bySlot: Map<DerivedArmorPieceJson["slot"], DerivedArmorPieceJson[]>,
-): Record<(typeof ARMOR_STAT_NAMES)[number], number> {
-  const maxima = Object.fromEntries(
-    ARMOR_STAT_NAMES.map((stat) => [stat, 0]),
-  ) as Record<(typeof ARMOR_STAT_NAMES)[number], number>;
-  for (const slot of SLOT_ORDER) {
-    for (const piece of bySlot.get(slot) ?? []) {
-      for (const stat of ARMOR_STAT_NAMES) {
-        maxima[stat] = Math.max(maxima[stat], getPieceStatCeiling(piece, stat));
-      }
+  let total = productSlotCounts(legendariesBySlot);
+
+  for (const exoticSlot of SLOT_ORDER) {
+    const exotics = (bySlot.get(exoticSlot) ?? []).filter((piece) => piece.isExotic);
+    const exoticCount = dedupedRepresentativeCount(exotics);
+    if (exoticCount === 0) continue;
+
+    let otherProduct = 1;
+    for (const slot of SLOT_ORDER) {
+      if (slot === exoticSlot) continue;
+      const count = dedupedRepresentativeCount(
+        legendariesBySlot.get(slot) ?? [],
+      );
+      if (count === 0) return total;
+      otherProduct *= count;
     }
+    total += exoticCount * otherProduct;
   }
-  return maxima;
+
+  return total;
 }
 
 export type FilteredComboCountOptions = {
@@ -95,34 +136,19 @@ export function estimateFilteredComboCount(
   const fragmentOffset = options.statOffset ?? {};
   const cap = options.cap;
 
-  if (
-    !hasStatTargets(constraints) &&
-    setBonusSelections.length === 0
-  ) {
+  if (!hasStatTargets(constraints) && setBonusSelections.length === 0) {
     return {
       count: estimateOptimizerComboCount(pool, exoticLock),
       capped: false,
     };
   }
 
-  const bySlot = groupPoolBySlot(pool);
-  const lockedIdentityKey = resolveLockedExoticIdentityKey(exoticLock, pool);
-  applyExoticLockToSlotGroups(bySlot, exoticLock, pool);
-
-  for (const slot of SLOT_ORDER) {
-    if ((bySlot.get(slot)?.length ?? 0) === 0) {
-      return { count: 0, capped: false };
-    }
-    const { representatives } = dedupeSlotPieces(bySlot.get(slot) ?? []);
-    bySlot.set(slot, representatives);
+  const prepared = prepareDedupedSlotPool({ pool, exoticLock });
+  if (prepared == null) {
+    return { count: 0, capped: false };
   }
 
-  const perSlotMax = perSlotMaxima(bySlot);
-  const slotPieces = SLOT_ORDER.map((slot) => bySlot.get(slot) ?? []);
   const verifyAtLeaves = hasStatTargets(constraints);
-  let count = 0;
-  let capped = false;
-
   const zeroTotals = totalsFromPieces([]);
   const startTotals =
     Object.keys(fragmentOffset).length > 0
@@ -132,19 +158,17 @@ export function estimateFilteredComboCount(
         )
       : zeroTotals;
 
-  const visit = (
-    slotIndex: number,
-    chosen: DerivedArmorPieceJson[],
-    partialTotals: Record<(typeof ARMOR_STAT_NAMES)[number], number>,
-  ): void => {
-    if (capped) return;
-
-    if (slotIndex >= SLOT_ORDER.length) {
-      if (exoticLock.mode === "any" && countExoticsInPieces(chosen) > 1) {
-        return;
-      }
+  const { leafCount, capped } = enumerateLoadouts({
+    prepared,
+    exoticLock,
+    startTotals,
+    constraints,
+    assumedMods,
+    setBonusSelections,
+    cap,
+    onLeaf: (chosen) => {
       if (!satisfiesSetBonuses(chosen, setBonusSelections)) {
-        return;
+        return "reject";
       }
       if (verifyAtLeaves) {
         const resolved = resolveLoadoutTotals(
@@ -154,61 +178,14 @@ export function estimateFilteredComboCount(
           assumedMods,
         );
         if (resolved == null) {
-          return;
+          return "reject";
         }
       }
-      count += 1;
-      if (cap != null && count >= cap) {
-        capped = true;
-      }
-      return;
-    }
+      return "accept";
+    },
+  });
 
-    const remaining = SLOT_ORDER.length - slotIndex;
-    const remainingSlots = SLOT_ORDER.slice(slotIndex);
-    for (const piece of slotPieces[slotIndex] ?? []) {
-      if (
-        !exoticAllowedInPartialCombo(
-          piece,
-          chosen,
-          exoticLock,
-          lockedIdentityKey,
-        )
-      ) {
-        continue;
-      }
-      const nextTotals = { ...partialTotals };
-      for (const stat of ARMOR_STAT_NAMES) {
-        nextTotals[stat] += getPieceStatCeiling(piece, stat);
-      }
-      if (
-        !partialCanReachMins(
-          nextTotals,
-          remaining - 1,
-          perSlotMax,
-          constraints,
-          assumedMods,
-        )
-      ) {
-        continue;
-      }
-      if (
-        !partialCanSatisfySetBonuses(
-          [...chosen, piece],
-          remainingSlots.slice(1),
-          bySlot,
-          setBonusSelections,
-        )
-      ) {
-        continue;
-      }
-      visit(slotIndex + 1, [...chosen, piece], nextTotals);
-      if (capped) return;
-    }
-  };
-
-  visit(0, [], startTotals);
-  return { count, capped };
+  return { count: leafCount, capped };
 }
 
 export type StatTargetFeasibilityOptions = Omit<
@@ -241,9 +218,7 @@ export function maxFeasibleStatTarget(
   focusStat: ArmorStatName,
   options: StatTargetFeasibilityOptions = {},
 ): number {
-  const hi = clampOptimizerStat(
-    options.hi ?? OPTIMIZER_STAT_MAX,
-  );
+  const hi = clampOptimizerStat(options.hi ?? OPTIMIZER_STAT_MAX);
   const feasible = (min: number): boolean =>
     estimateFilteredComboCount(pool, exoticLock, {
       ...options,
