+++ src/lib/optimizer/constants.ts
@@ -1,0 +1,5 @@
+/** Joint slider bands disabled on the hot path (see useStatBoundsForSliders). */
+export const JOINT_BOUNDS_COMBO_LIMIT = 0;
+
+/** Above this deduped combo count, auto-search is skipped (manual run). */
+export const SEARCH_AUTO_RUN_COMBO_LIMIT = 50_000;
